+++ src/test/java/org/apiphany/security/ssl/client/AlertLevel.java
@@ -24,7 +24,7 @@ public enum AlertLevel implements TLSObject {
 	}
 
 	@Override
-	public int size() {
+	public int sizeOf() {
 		return BYTES;
 	}