+++ src/test/java/org/apiphany/security/ssl/client/CipherSuite.java
@@ -110,7 +110,7 @@ public enum CipherSuite implements TLSObject {
 	}
 
 	@Override
-	public int size() {
+	public int sizeOf() {
 		return BYTES;
 	}