+++ src/test/java/org/apiphany/security/ssl/client/ECPointFormats.java
@@ -50,7 +50,7 @@ public class ECPointFormats implements TLSExtension {
 
 	@Override
 	public byte[] toByteArray() {
-		ByteBuffer buffer = ByteBuffer.allocate(size());
+		ByteBuffer buffer = ByteBuffer.allocate(sizeOf());
 		buffer.put(type.toByteArray());
 		buffer.put(length.toByteArray());
 		buffer.put(listSize.toByteArray());
@@ -66,8 +66,8 @@ public class ECPointFormats implements TLSExtension {
 	}
 
 	@Override
-	public int size() {
-		return type.size() + length.size() + listSize.size() + formats.size();
+	public int sizeOf() {
+		return type.sizeOf() + length.sizeOf() + listSize.sizeOf() + formats.size();
 	}
 
 	@Override