+++ src/test/java/org/apiphany/security/ssl/client/ExtensionType.java
@@ -49,7 +49,7 @@ public enum ExtensionType implements TLSObject {
 	}
 
 	@Override
-	public int size() {
+	public int sizeOf() {
 		return BYTES;
 	}