+++ src/test/java/org/apiphany/security/ssl/client/PublicKeyRSA.java
@@ -34,7 +34,7 @@ public class PublicKeyRSA implements TLSObject {
 
 	@Override
 	public byte[] toByteArray() {
-		ByteBuffer buffer = ByteBuffer.allocate(size());
+		ByteBuffer buffer = ByteBuffer.allocate(sizeOf());
 		buffer.put(length.toByteArray());
 		buffer.put(bytes.toByteArray());
 		return buffer.array();
@@ -54,8 +54,8 @@ public class PublicKeyRSA implements TLSObject {
 	}
 
 	@Override
-	public int size() {
-		return length.size() + bytes.size();
+	public int sizeOf() {
+		return length.sizeOf() + bytes.sizeOf();
 	}
 
 }