+++ src/test/java/org/apiphany/security/ssl/client/RecordContentType.java
@@ -44,7 +44,7 @@ public enum RecordContentType implements TLSObject {
 	}
 
 	@Override
-	public int size() {
+	public int sizeOf() {
 		return BYTES;
 	}