+++ src/test/java/org/apiphany/security/ssl/client/SessionTicket.java
@@ -53,8 +53,8 @@ public class SessionTicket implements TLSExtension {
 	}
 
 	@Override
-	public int size() {
-		return type.size() + length.size();
+	public int sizeOf() {
+		return type.sizeOf() + length.sizeOf();
 	}
 
 	@Override